move dismiss service to html5
